Refactor Tensor::to to call a primitive that is not copy_.
functorch is unable to vmap(grad(f)) when f contains a .to
call. This is because .to (when it is not a no-op) decomposes
to .copy_ under grad and the .copy_ is not compatible with vmap.

The fix for this is to have all Tensor::to variants call a new operator,
`_to_copy`, that always copies and is a primitive w.r.t. autograd so
that autograd decomposes Tensor::to into a call to `_to_copy`.
(This is related to https://github.com/pytorch/pytorch/issues/60956,
please let me know if you want to bikeshed the naming).

In order to get this done I had to do a bit of refactoring. All of the
`::to` implementations now call `to_impl` which may call `_to_copy`.

Performance benchmarking:
- I benchmarked [three
cases](https://gist.github.com/zou3519/5b6985e6906b80eec5a0dd94ed5b6a1a).
- Case A: No-op .to(). Instruction count went from 49923 to 25123. I
have no clue why but this is a good thing.
- Case B: not-no-op .to(). Instruction count went from 662816 to 671649.
This is expected; `_to_copy` adds an additional dispatch.
- Case C: not-no-op .to() forward pass and backward pass. Instruction count
went from 4022752 to 4067535. This is a little concerning but I don't
have a good answer for why. This PR adds
an additional dispatch to .to() (so there should be one additional
dispatch in the forward pass and one additional dispatch in the backward
pass). It might just be that the Autograd Node creation for `copy_` is
really fast.

Test Plan:
- test_torch.py has a test_to
- test_cuda.py has test_to*
- In the near future we will set up a functorch CI in PyTorch CI

ghstack-source-id: dc4403d98cfe55b27fe2aedd0a298291ba143d24
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61458